Reapply "Remove requirement that fsv_meta file is next to its inp..."
In soong-built partitions, we will generate fsv_meta files before
a partition staging directory is generated. This means that the source
file will be some module-specific intermediates path, and the dest
file will be a path local to the filesystem module. Relax the
requirement that fsv_meta files are next to the files they protect.

Bug: 394404628
Test: m aosp_shared_system_image, verified it didn't change before/after
This reverts commit 0b63a6e0b1412108eee6e9ce38678215ae43c623.

Change-Id: I69367e599f71c6ac4c7ce0bcd4a5aaead6ec6d9b