Remove redundant entries from symbols.zip
Currently, symbols.zip contains multiple duplicates of the symbols file,
which are derived from uninstalled soong modules that produce androidmk
entries and share the identical symbols file with the installed modules.
These are likely unintended behavior, given that the current build rule
of the symbols.zip creates dependency on all generated images instead of
all modules that generate the symbols file.

This change modifies the build rule of the symbols.zip file so that
only the symbols files of the installed modules or the dependencies of
the test suites are included in the zip file.

Test: m droid dist && inspect the entries included in the symbols.zip file
Bug: 395989947
Change-Id: Icd55ed1bd76b688829f302daebbd1cc32356fe30