aconfig: Move SDK check within Java codegen
Having it within the Flags class makes it easier to remove the "else"
branch for binaries with high enough minSdk. Most if not all uses are
here rather than within FeatureFlags, but left the check in
FeatureFlagsImpl just to be safe.

Also moved within ExportedFlags which is the codegen for apps to use.

Bug: 378936061
Test: manually built the files, atest aconfig.test

Change-Id: Ib487503970a046e913977093fe09c1f7adb96e41