add tool exported-flag-check
This tool ensures that exported flags are used as intended. Exported flags, marked with
`is_exported: true` in their declaration, are designed to control access to specific API
features. This tool identifies and reports any exported flags that are not currently
associated with an API feature, preventing unnecessary flag proliferation and maintaining
a clear API design.

This tool should be triggered by the build system when it builds
all_aconfig_declarations

Test: atest exported-flag-check
Bug: 336800305
Change-Id: Ia1574cb05478eb2f329b959ee69ad13a4abdad50